+++ .projenrc.ts
@@ -61,6 +61,6 @@ const project = new awscdk.AwsCdkConstructLibrary({
   // packageName: undefined,  /* The "name" in package.json. */
 });
 project.setScript('cdk', 'cdk');
-project.setScript('integ:deploy', "cdk deploy --app='./lib/integ.js'");
-project.setScript('integ:destroy', "cdk destroy --app='./lib/integ.js'");
+project.setScript('integ:deploy', "cdk deploy --app='./test/integ.js'");
+project.setScript('integ:destroy', "cdk destroy --app='./test/integ.js'");
 project.synth();
